Crate moore_svlog

This crate implements SystemVerilog for the moore compiler.

Reexports

pub extern crate moore_svlog_syntax;

pub extern crate moore_svlog_hir;

pub use moore_svlog_syntax::*;

Modules

hir
The high-level intermediate representation for SystemVerilog. After name resolution, the AST is lowered into this representation, eliminating a lot of the syntactic sugar represented in the AST, and resolving default and implicit values.