Module moore_svlog::value

Representation of constant values and their operations

This module implements a representation for values that may arise within a SystemVerilog source text and provides ways of executing common operations such as addition and multiplication. It also provides the ability to evaluate the constant value of nodes in a context.

The operations in this module are intended to panic if invalid combinations of values are used. The compiler's type system should catch and prevent such uses.

Structs

ValueData
The data associated with a value.

Enums

ValueKind
The different forms a value can assume.

Functions

make_array
Create a new array value.

make_int
Create a new integer value.

make_int_special
Create a new integer value with special bits.

make_struct
Create a new struct value.

make_time
Create a new time value.

Type Definitions

Value
A verilog value.