Struct moore_svlog::GlobalContext

pub struct GlobalContext<'gcx> {
    pub sess: &'gcx Session,
    pub arena: &'gcx GlobalArenas<'gcx>,
    // some fields omitted
}

The central data structure of the compiler. It stores references to various arenas and tables that store the results of the various computations that have been performed.

Fields
sess: &'gcx Session

The global compiler session.

arena: &'gcx GlobalArenas<'gcx>

The arena that owns all references.

Methods

impl<'gcx> GlobalContext<'gcx>

pub fn new(sess: &'gcx Session, arena: &'gcx GlobalArenas<'gcx>) -> Self

Create a new global context.

pub fn add_root_nodes(
&self,
ast: impl Iterator<Item = &'gcx Root>
)

Add root AST nodes to the context for processing.

Use the find_global_item function afterwards to look up the id of modules that were added.

pub fn find_module(&self, name: Name) -> Option<NodeId>

Find a module in the AST.

pub fn modules(
&self
) -> impl Iterator<Item = (Name, NodeId)>

Get an iterator over all modules in the AST.