Trait moore_svlog::Context

pub trait Context<'a>: GetQueryTable<HirOfQuery<'a>> + GetQueryTable<ParamEnvQuery<'a>> + GetQueryTable<TypeOfQuery<'a>> + GetQueryTable<MapToTypeQuery<'a>> + GetQueryTable<LocalRibQuery<'a>> + GetQueryTable<ResolveUpwardsQuery<'a>> + GetQueryTable<ResolveNodeQuery<'a>> + GetQueryTable<ConstantValueOfQuery<'a>> + GetQueryTable<IsConstantQuery<'a>> + GetQueryTable<TypeDefaultValueQuery<'a>> + GetQueryTable<AccessedNodesQuery<'a>> + GetQueryTable<PortMappingQuery<'a>> + GetQueryTable<ResolveFieldAccessQuery<'a>> + BaseContext<'a> {
    fn hir_of(&self, node_id: NodeId) -> Result<HirNode<'a>> { ... }

    fn param_env(&self, src: ParamEnvSource<'a>) -> Result<ParamEnv> { ... }

    fn type_of(&self, node_id: NodeId, env: ParamEnv) -> Result<Type<'a>> { ... }

    fn map_to_type(&self, node_id: NodeId, env: ParamEnv) -> Result<Type<'a>> { ... }

    fn local_rib(&self, node_id: NodeId) -> Result<&'a Rib> { ... }

    fn resolve_upwards(
        &self, 
        name: Name, 
        start_at: NodeId
    ) -> Result<Option<NodeId>> { ... }

    fn resolve_node(&self, node_id: NodeId, env: ParamEnv) -> Result<NodeId> { ... }

    fn constant_value_of(
        &self, 
        node_id: NodeId, 
        env: ParamEnv
    ) -> Result<Value<'a>> { ... }

    fn is_constant(&self, node_id: NodeId) -> Result<bool> { ... }

    fn type_default_value(&self, ty: Type<'a>) -> Value<'a> { ... }

    fn accessed_nodes(&self, node_id: NodeId) -> Result<Arc<AccessTable>> { ... }

    fn port_mapping(
        &self, 
        src: PortMappingSource<'a>
    ) -> Result<Arc<PortMapping>> { ... }

    fn resolve_field_access(
        &self, 
        node_id: NodeId, 
        env: ParamEnv
    ) -> Result<(NodeId, usize, NodeId)> { ... }
}

A collection of compiler queries.

Provided Methods

fn hir_of(&self, node_id: NodeId) -> Result<HirNode<'a>>

Lower an AST node to HIR.

fn param_env(&self, src: ParamEnvSource<'a>) -> Result<ParamEnv>

Compute the parameter bindings for an instantiation.

fn type_of(&self, node_id: NodeId, env: ParamEnv) -> Result<Type<'a>>

Determine the type of a node.

fn map_to_type(&self, node_id: NodeId, env: ParamEnv) -> Result<Type<'a>>

Convert a node to a type.

fn local_rib(&self, node_id: NodeId) -> Result<&'a Rib>

Determine the local rib that applies to a node.

fn resolve_upwards(
&self,
name: Name,
start_at: NodeId
) -> Result<Option<NodeId>>

Resolve a name upwards through the ribs.

fn resolve_node(&self, node_id: NodeId, env: ParamEnv) -> Result<NodeId>

Resolve a node to its target.

fn constant_value_of(&self, node_id: NodeId, env: ParamEnv) -> Result<Value<'a>>

Determine the constant value of a node.

fn is_constant(&self, node_id: NodeId) -> Result<bool>

Check if a node has a constant value.

fn type_default_value(&self, ty: Type<'a>) -> Value<'a>

Determine the default value of a type.

fn accessed_nodes(&self, node_id: NodeId) -> Result<Arc<AccessTable>>

Determine the nodes accessed by another node.

fn port_mapping(&self, src: PortMappingSource<'a>) -> Result<Arc<PortMapping>>

Compute the port assignments for an instantiation.

fn resolve_field_access(
&self,
node_id: NodeId,
env: ParamEnv
) -> Result<(NodeId, usize, NodeId)>

Resolve the field name in a field access expression.

Returns the node id of the corresponding struct definition, the index of the field that is actually being accessed, and the node id of the field definition.